move router common tests to scoped koa router test file
